NEW VERSION v3.0
This version feature a complete treatmente of the inequivalent sites and other inhomogeneous geometries (RDMFT).
The structure of the code is such to preserve the original version, dealing with the single site problem.
Around these original modules suitable WRAPPERS have been introduced (ED_WRAP_xxx) that extend the functionalities
to the case of inequivalent sites.

News:
- introduced new modules to evaluate the local Green's functions and Weiss Field update (DMFT self-consistency)
ED_GLOC.f90, ED_WEISS.f90
- big update of the ED_MAIN.f90 module which offer the common interface to the user for the single site problem.
- introduced a set of routines to retrieve several quantities from the code, without accessing the memory (in ED_MAIN,
ED_WRAP_MAIN), such as get the self-energies, get the impurity green's function, get the densities, double occupancies,
potential and local energies.
- introduced new wrappers to the fundamental interface routines to the inequivalent sites problem.
ED_MAIN --> ED_WRAP_MAIN
ED_GLOC --> ED_WRAP_GLOC
ED_WEISS --> ED_WRAP_WEISS
ED_ENERGY --> ED_WRAP_ENERGY
ED_CHI2FIT --> ED_WRAP_CHI2FIT

plus a ED_WRAP_AUX_FUNX which contains the necessary auxiliary routines for the wrappers.

The fundamental module is DMFT_ED which expose all the necessary routines to the user.

TODO:
- there is obviously a huge problem with the parallelization of the codes now.
in practice we have two possible setup now: parallelize with respect to the inequivalent sites (RDMFT-style) (mpiID),
parallelize with respect to the ED code (ED_MPI_ID).  In the future the plan is to use both setup by creating on the
fly new communicators, but one has to pick up the best possible strategy for this that allows to do heavy parallel
ED calculation for one site, or large number of ineq. sites.
- extend the GLOC and WEISS modules to the case of p-d problems including orbitals which are not explicitly treated
in the solver (non-correlated electrons). This is relatively easy to achieve, introducing a total number of orbitals
Ntorb = Norb + Nporb.

TEST:
- tests are currently running now. First results obtained with the attractive single band on the square lattice gave
nice results, though a different convergence behavior was observed between Intel compiler and GNU compiler, final results
are anyway nicely in agreement.
UPDATE TEST:
tests have been performed in many different cases: homo attractive hubbard model, inhomo attractive hubbard, homo BHZ, inhomo
2 bands with hybridization slab geometry. All gave exact results against those obtained with previous code.

+++ PLAIN_LANCZOS.f90
@@ -28,7 +28,7 @@ MODULE PLAIN_LANCZOS
   procedure(lanc_htimesv_c),pointer     :: cp_hprod
   logical                               :: verb=.false.
   real(8)                               :: threshold_=1.d-15
-  real(8)                               :: ncheck_=10
+  integer                               :: ncheck_=10
 
   public :: lanczos_plain_iteration_d
   public :: lanczos_plain_iteration_c
